feat: hooks that you can use to create custom wallet connection UI components (#789)
* feat: hooks that you can use to create custom wallet connection UI components
## Summary

In this PR we create hooks that track the state of the wallet connection specifically for the purpose of rendering wallet connection UI. This will allow UI developers to create custom controls easily, using their own UI frameworks, localization infrastructure, and styles.

In the next PR we rebuild the React, Material UI, and Ant Design libraries atop this hook. This gives three concrete examples of how it can be used to build custom UI with a minimum of effort.

## Test Plan

See next PR where we rebuild the React, Material UI, and Ant Design starters atop this hook.

Implements #658.